+++ mkapi/core/postprocess.py
@@ -11,8 +11,8 @@ def sourcelink(object: Object) -> str:
         link = f'<span id="{object.id}"></span>'
     else:
         link = ""
-    link += f'<a class="mkapi-args-src-link" href="../source/{object.module}'
-    link += f'#{object.id}" title="Source for {object.id}"></a>'
+    link += f'<a class="mkapi-item-src-link" href="../source/{object.module}'
+    link += f'#{object.id}" title="Source for {object.id}">&lt;/&gt;</a>'
     return link